Rethink JWT authentication
