+++ src/features/home/homePage.tsx
@@ -1,5 +1,4 @@
 import type { AppVariables } from "@shared/appVariables.ts";
-import { parseJwtPayload } from "@shared/jwtPayload.ts";
 import { Page } from "@shared/page.tsx";
 import { pageJwtMiddleware } from "@shared/pageJwtMiddleware.ts";
 import { Hono } from "hono";
@@ -7,7 +6,7 @@ import { Hono } from "hono";
 export const homePage = new Hono<{ Variables: AppVariables }>()
   .use("/", pageJwtMiddleware)
   .get("/", (c) => {
-    const username = parseJwtPayload(c.var.jwtPayload).preferred_username;
+    const username = c.var.validJwtPayload.preferred_username;
     return c.html(
       <Page type="authenticated" currentPath={c.req.path} username={username}>
         <h1>Home</h1>
