+++ src/features/login/loginApi.tsx
@@ -1,7 +1,7 @@
 import { zValidator } from "@hono/zod-validator";
 import type { AppVariables } from "@shared/appVariables.ts";
 import { sseRedirect } from "@shared/datastar.ts";
-import { createTodoJwtPayload } from "@shared/jwtPayload.ts";
+import { createTodoJwtPayload } from "@shared/pageJwtMiddleware.ts";
 import { pageRoutes } from "@shared/pageRoutes.ts";
 import { getUserByUsername } from "@shared/user.ts";
 import { addDays } from "date-fns";
